requirements.txt
Flask
qrcode
Pillow
pyzbar
requests
gunicorn
python-dotenv
